refactor(util): random package를 추가한다
random package를 추가하여 package 간 책임을 분리한다.
